+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ## [Unreleased](https://gitlab.heigit.org/climate-action/plugins/walkability/-/compare/4.2.1...main)
 
+### Changed
+- Color categories individually in POI summary chart ([#390](https://gitlab.heigit.org/climate-action/plugins/walkability/-/work_items/390))
+
 ## [4.2.1](https://gitlab.heigit.org/climate-action/plugins/walkability/-/releases/4.2.1) - 2026-08-18
 
 ### Added